get rid of unnecessary abstraction and rename types accordingly
